Remove patch functionality
While really neat in theory, for me patching isn't useful. I'd much prefer
to be able to edit my files in place, then go to my dotfile and make a commit.
The patch workflow got in the way of that by making the link between the
generated file instead of the src file.

I also think this works better for me because I'd rather my dotfiles handle
working on different platforms instead of patching them to work on different
platforms.